refactor Response add Session Object
This is partial refactor.  ArubaCentralBase will be replaced
by Session object and take over everything it is doing now.
That library has yet to be updated since it's posting and we are
overriding some of its methods anyway.

`self.central` will eventually become the object that represents
the session with Aruba Central
What is currently central.py will be broken into
monitoring.py, inventory.py, etc