Change config dir logic
Found on windows with pypi install that
typer.get_app_dir(__name__) returned the full path
in site_packages, not what we want.

program now looks for config in a number of places:
```
            [
                Path().home() / ".config" / "centralcli",
                Path().home() / ".centralcli",
                cwd / "config",
                cwd,
                Path().home() / ".config" / "centralcli" / "config",
            ]
```
If the config file is in any of those then that's what is
used, and directory structure is derived from there.