enhance batch add sites
Made it more flexible in regard to the format of the input file

site_name can be site_name, name, or site
address and geoloc fields can but don't have to be under the site_address
or geolocation keys.

The command will catch *most* / not all imput data file errors.

Does not verify input data where address/geo are already in site_address
/ geolocation key.