enhance completions add `test method` completion dev/group/site completion enhanced so that when multiples are allowed completion does not return items already on the command line.
hook completion was added for events.  webhook id to webhook name not in
place yet, but cache is ready for it.

sites completion fixed to handle no value "" (returns all sites not on CLI yet)
Due to it being valid that some fields would have no value completion
was not behaving as expected for sites unless the word was started.