always set LESS environ var
Prior I only modified if it wasn't already set.  It only persists
for the duration of the command.

default will be not to page, given most of us don't like the paging behavior
and you can pipe to a pager of your liking in the shell... Not implemented yet tho