proper per second rate limit
hours of work, most of which is essentially removed, replaced
with the logic in _batch requests which does 7 calls at a time
if they finish faster than 1 second, it will wait the difference beforfe
sending the next chunk.  Could not get anything else to work all the time
this logic seems to completely prevent an overrun.