honor `stash` parameter in display_results()
stash controls if the output of a command is cached for re-display with
`cencli show last`.  It allows you to re-display the output without
making an API call.

Handy if you want to see the same results in a different format
or decide you want to dump the results to a file.