return unaltered payload when payload doesn't look like it's wrapped
`strip_outer_keys()` removes the parent key which varries depending on endpoint
for example get aps will return a dict with {"aps", <lisf of dicts>, count: int}
or {"clients": <list of dicts>, count: int, last_client_mac: str]
etc.

This change resolves the issue where the API was querried for details of a specific
client (by mac) a dict was returned with no wrapping key, but this method
was seeing the "labels" key in the payload and returning that value.

Given a wrapping key scenario is expected to have at most 3 keys, `> 5` should handle it.
(Not sure if there is a scenario where it would have more than 3 hence 5
for some margin of safety.