fix issue when cleaner called x2 for same data
Changed device cache update so it keeps the formatting of the response
output unmodified.  vs updating output attr with clean values.

Issue was cleaner would be called a second time on already formatted data
which is an issue in some scenarios

i.e. uptime already converted to words when its passed through a cleaner
func to convert an int to words.