:bookmark: v2.3.3 :sparkles: Add commands to show/disable/enable auto-subscription
:bug: The enable/disable commands are currently hidden.
As GreenLake and the API have some inconsistencies.

auto-sub advanced-switch-6300 will result in the API showing it is
enabled for all switches (this is OK).   However GreenLake doesn't reflect
the change.  The API to disable unsubscribe will return OK, but doesn't appear
to change anything.  The only way to get the API (that shows what is enabled)
to clear auto-sub is to create then delete auto-sub from GreenLake UI.