🧑‍💻 Add `caption` attribute to Response object, Add `has_value` property to RateLimit object.
caption attribute can be used to update captions using the resp object.  has_value is used to determine if headers were sent with rate limit details.  Rate limit is not added to the caption if no details were extracted from response headers... Which is the case fot glp/cnx APIs.