🔊 capture_logs... add expect_failure parmeter for capturing output when we expect return code of 1
When we initiate the exit with an error code it's always error_code=1, typer will always exit with error_code 2+, so we can still evaluate if the command exits as expected even if we expect it to fail.  This also has the benefit of grabbing any exception details if the command bombs.