🚚 get_dev_identifier.  rename params conductor_only is now swack.  swack is not swack_only. 🐛 fix UnboundLocalError on multi-match with conductor_only option (now swack) in get_dev_identifier
swack: return any matches, but if there are multiple matches and any are members of a swarm or stack only return the VC/Conductor
swack_only: Same as above but any matches that are not members of a swarm or stack are filtered.

This seems more logical than what the params were, which I always had to go double check the original meaniing of.